pyproject.toml
[project]
name = "ai-resume-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "gradio>=5.42.0",
    "huggingface-hub>=0.34.4",
    "ipykernel>=6.30.1",
    "ipython>=9.4.0",
    "openai>=1.99.9",
    "pydantic>=2.11.7",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
